requirements.txt
# Main
RPi.GPIO
PyYAML
# Audio
aiohttp==0.16.2
asyncio==3.4.3
pyaudio==0.2.8
git+https://github.com/NuanceDev/pyspeex.git@0.9.0#egg=pyspeex
SpeechRecognition==3.4.6
pocketsphinx==0.1.3
agecalc==3.5.2
hug
